% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{duke}
\alias{duke}
\title{Duke breast cancer data}
\format{
A data frame with 86 rows and 501 variables,
where the first variable is the label of estrogen receptor-positive/negative,
and the remaining 500 variables are 500 gene.
}
\source{
\url{www.kaggle.com/andreicosma/duke-breast-cancer-dataset}
}
\description{
This data set details microarray experiment for breast cancer patients.
}
\details{
The binary variable Status is used to classify the patients into
estrogen receptor-positive (y = 0) and estrogen receptor-negative (y = 1).
The other variables contain the expression level of the considered genes.
}
\references{
M. West, C. Blanchette, H. Dressman, E. Huang, S. Ishida, R. Spang, H. Zuzan, J.A. Olson, Jr., J.R. Marks and Joseph R. Nevins (2001) <doi:10.1073/pnas.201162998> Predicting the clinical status of human breast cancer by using gene expression profiles, Proceedings of the National Accademy of Sciences of the USA, Vol 98(20), 11462-11467.
}
