% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.bess.R
\name{coef.bess}
\alias{coef.bess}
\title{Provides estimated coefficients from a fitted "bess" object.}
\usage{
\method{coef}{bess}(object, sparse = TRUE, ...)
}
\arguments{
\item{object}{A "\code{bess}" project.}

\item{sparse}{Logical or NULL, specifying whether the coefficients should be
presented as sparse matrix or not.}

\item{\dots}{Other arguments.}
}
\description{
Similar to other prediction methods, this function provides estimated
coefficients from a fitted "\code{bess}" object.
}
\examples{


# Generate simulated data
n = 200
p = 20
k = 5
rho = 0.4
SNR = 10
cortype = 1
seed = 10
Data = gen.data(n, p, k, rho, family = "gaussian", cortype = cortype, SNR = SNR, seed = seed)
x = Data$x[1:140, ]
y = Data$y[1:140]
x_new = Data$x[141:200, ]
y_new = Data$y[141:200]
lm.pdas = bess(x, y, method = "sequential")
lambda.list = exp(seq(log(5), log(0.1), length.out = 10))
lm.l0l2 = bess(x, y, type = "bsrr")

coef(lm.pdas)
coef(lm.l0l2, sparse = TRUE)




}
\references{
Wen, C., Zhang, A., Quan, S. and Wang, X. (2020). BeSS: An R
Package for Best Subset Selection in Linear, Logistic and Cox Proportional
Hazards Models, \emph{Journal of Statistical Software}, Vol. 94(4).
doi:10.18637/jss.v094.i04.
}
\seealso{
\code{\link{bess}}, \code{\link{print.bess}}.
}
\author{
Liyuan Hu, Aijun Zhang, Shijie Quan, and Xueqin Wang.
}
