% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bess.R
\name{plot.bess}
\alias{plot.bess}
\title{Produces a coefficient profile plot of the coefficient or loss function
paths}
\usage{
\method{plot}{bess}(
  x,
  type = c("loss", "tune", "coefficients", "both"),
  breaks = TRUE,
  K = NULL,
  sign.lambda = 0,
  ...
)
}
\arguments{
\item{x}{a "\code{bess}" project fitted via "PDAS" algorithm.}

\item{type}{Either "both", "solutionPath" or "loss".}

\item{breaks}{If TRUE, then vertical lines are drawn at each break point in
the coefficient paths}

\item{K}{which break point should the vertical lines drawn at}

\item{sign.lambda}{Whether to show lambda on log scale. Default is 0.}

\item{\dots}{Other graphical parameters to plot}
}
\description{
Produces a coefficient profile plot of the coefficient or loss function
paths
}
\examples{


#-------------------linear model----------------------#
# Generate simulated data
n = 200
p = 20
k = 5
rho = 0.4
SNR = 10
cortype = 1
seed = 10
Data = gen.data(n, p, k, rho, family = "gaussian", cortype=cortype, SNR=SNR, seed=seed)
x = Data$x[1:140, ]
y = Data$y[1:140]
x_new = Data$x[141:200, ]
y_new = Data$y[141:200]
lm.pdas = bess(x, y, method = "sequential")
lm.l0l2.powell = bess(x, y, type = "bsrr")
lm.l0l2.seq = bess(x, y, type = "bsrr", method = "psequential")

# Plot the solution path and the loss function of PDAS
plot(lm.pdas, type = "both", breaks = TRUE)

# Plot the solution path of the best ridge regreesion
plot(lm.l0l2.powell)

plot(lm.l0l2.seq)


}
\references{
Wen, C., Zhang, A., Quan, S. and Wang, X. (2020). BeSS: An R
Package for Best Subset Selection in Linear, Logistic and Cox Proportional
Hazards Models, \emph{Journal of Statistical Software}, Vol. 94(4).
doi:10.18637/jss.v094.i04.
}
\seealso{
\code{\link{bess}}.
}
\author{
Liyuan Hu.
}
