% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bess.R
\name{print.bess}
\alias{print.bess}
\title{print method for a "bess" object}
\usage{
\method{print}{bess}(x, digits = max(5, getOption("digits") - 5), nonzero = FALSE, ...)
}
\arguments{
\item{x}{A "\code{bess}" object.}

\item{digits}{Minimum number of significant digits to be used.}

\item{nonzero}{Whether the output should only contain the non-zero coefficients.}

\item{\dots}{additional print arguments}
}
\description{
Print the primary elements of the "\code{bess}" object.
}
\details{
prints the fitted model and returns it invisibly.
}
\examples{


# Generate simulated data
n = 200
p = 20
k = 5
rho = 0.4
SNR = 10
cortype = 1
seed = 10
Data = gen.data(n, p, k, rho, family = "gaussian", cortype=cortype, snr=SNR, seed=seed)
x = Data$x[1:140, ]
y = Data$y[1:140]
x_new = Data$x[141:200, ]
y_new = Data$y[141:200]
lm.bss = bess(x, y, method = "sequential")
lambda.list = exp(seq(log(5), log(0.1), length.out = 10))
lm.bsrr = bess(x, y, type = "bsrr", lambda.list = lambda.list, method = "sequential")

print(lm.bss)
print(lm.bsrr)

}
\references{
Wen, C., Zhang, A., Quan, S. and Wang, X. (2020). BeSS: An R
Package for Best Subset Selection in Linear, Logistic and Cox Proportional
Hazards Models, \emph{Journal of Statistical Software}, Vol. 94(4).
doi:10.18637/jss.v094.i04.
}
\seealso{
\code{\link{bess}}, \code{\link{plot.bess}},
\code{\link{summary.bess}}.
}
\author{
Canhong Wen, Aijun Zhang, Shijie Quan, Liyuan Hu, Kangkang Jiang, Yanhang Zhang, Jin Zhu and Xueqin Wang.
}
