% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bess.R
\name{summary.bess}
\alias{summary.bess}
\title{summary method for a "bess.one" object}
\usage{
\method{summary}{bess}(object, ...)
}
\arguments{
\item{object}{a "bess" object}

\item{\dots}{additional print arguments}
}
\description{
Print a summary of the "bess.one" object.
}
\examples{


#-------------------linear model----------------------#
# Generate simulated data
n = 200
p = 20
k = 5
rho = 0.4
SNR = 10
cortype = 1
seed = 10
Data = gen.data(n, p, k, rho, family = "gaussian", cortype=cortype, SNR=SNR, seed=seed)
x = Data$x[1:140, ]
y = Data$y[1:140]
x_new = Data$x[141:200, ]
y_new = Data$y[141:200]
lm.pdas = bess(x, y)
lambda.list = exp(seq(log(5), log(0.1), length.out = 10))
lm.l0l2 = bess(x, y, type = "bsrr", lambda.list = lambda.list)
summary(lm.pdas)
summary(lm.l0l2)



}
\references{
Wen, C., Zhang, A., Quan, S. and Wang, X. (2020). BeSS: An R
Package for Best Subset Selection in Linear, Logistic and Cox Proportional
Hazards Models, \emph{Journal of Statistical Software}, Vol. 94(4).
doi:10.18637/jss.v094.i04.
}
\seealso{
\code{\link{bess}}.
}
\author{
Liyuan Hu.
}
